requirements.txt
absl-py==2.3.1
astunparse==1.6.3
attrs==20.3.0
autopep8==2.3.1
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
check50==3.3.11
clang-format==19.1.1
click==8.3.0
cryptography==43.0.0
EditorConfig==0.12.4
filelock==3.20.0
flatbuffers==25.2.10
fsspec==2025.9.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.1
h5py==3.14.0
hf-xet==1.1.10
huggingface-hub==0.35.3
icdiff==2.0.7
idna==3.8
jellyfish==0.11.2
Jinja2==3.1.4
joblib==1.4.2
jsbeautifier==1.15.1
lib50==3.0.11
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.1.0
networkx==3.5
nltk==3.9.1
numpy==2.1.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.16.0
packaging==24.1
pexpect==4.9.0
pillow==12.0.0
protobuf==5.29.5
ptyprocess==0.7.0
pycodestyle==2.12.0
pycparser==2.22
pygame==2.6.0
Pygments==2.19.2
python-magic==0.4.27
PyYAML==6.0.2
regex==2025.9.18
requests==2.32.3
rich==14.0.0
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.15.2
setuptools==74.0.0
six==1.16.0
soupsieve==2.6
style50==2.10.4
submit50==3.1.5
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==1.1.0
tf_keras==2.20.1
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
triton==3.5.0
typing_extensions==4.14.1
urllib3==2.2.2
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
